requirements.txt
absl-py==2.1.0
accelerate==0.27.2
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.6.0
anthropic==0.19.1
antlr4-python3-runtime==4.8
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
bleach==6.1.0
blessed==1.20.0
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
dacite==1.8.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.25.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.4.0
eval-type-backport==0.1.3
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fschat==0.2.36
fsspec==2024.2.0
fuzzywuzzy==0.18.0
gast==0.5.4
gcsfs==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gpustat==1.1
gradio==4.25.0
gradio_client==0.15.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
higher==0.2.1
hjson==3.1.0
htmlmin==0.1.12
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hydra-core==1.1.1
idna==3.6
ImageHash==4.3.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jaxtyping==0.2.33
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
json5==0.9.22
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
Levenshtein==0.25.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mizani==0.11.2
modelscope==1.13.3
mpmath==1.3.0
ms-swift==1.7.3
multidict==6.0.5
multimethod==1.11.2
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.6.5
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
openai==1.14.1
opencv-python==4.8.0.76
optimum==1.18.0
orjson==3.10.0
oss2==2.18.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
peft==0.9.0
pexpect==4.9.0
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
plotnine==0.13.5
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pyvene==0.1.1
PyWavelets==1.6.0
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.6
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.5
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.11.4
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==1.40.6
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.12
shtab==1.7.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.2
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.0
threadpoolctl==3.3.0
tiktoken==0.6.0
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.40.0
transformers-stream-generator==0.0.5
triton==2.2.0
trl==0.7.6
typeguard==2.13.3
typer==0.12.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tyro==0.8.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.27.1
visions==0.7.6
wandb==0.16.4
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wordcloud==1.9.3
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
ydata-profiling==4.7.0
zipp==3.17.0
